Update README.md
The project is implemented using the Python programming language. Python provides a wide range of libraries and modules that are well-suited for web scraping tasks.

BeautifulSoup is a Python library that helps parse and navigate HTML and XML documents. It is used in this project to extract specific elements and data from the HTML structure of Wikipedia pages.